--- src/main/java/enums/Department.java
@@ -1,9 +1,0 @@
-package enums;
-
-public enum Department {
-    FINANCE,         // Billing, accounts, and financial management.
-    RECORDS,         // Medical records and health information management.
-    PHARMACY,
-    WARD
-}
-//setters
