--- src/main/java/enums/Gender.java
@@ -1,7 +1,0 @@
-package enums;
-
-public enum Gender {
-    MALE,
-    FEMALE
-
-}
