--- src/main/java/enums/Qualification.java
@@ -1,8 +1,0 @@
-package enums;
-
-public enum Qualification {
-    MBBS,            // Bachelor of Medicine, Bachelor of Surgery
-    MD,              // Doctor of Medicine (postgraduate)
-    MS,              // Master of Surgery (postgraduate)
-    BDS           // Bachelor of Dental Surgery
-}
