--- src/main/java/enums/Rolename.java
@@ -1,9 +1,0 @@
-package enums;
-
-public enum Rolename {
-    DOCTOR,
-    PHARMACIST,
-    RECEPTIONIST,
-    ADMINISTRATOR,
-    PATIENT
-}
