--- src/main/java/enums/Shift.java
@@ -1,13 +1,0 @@
-package enums;
-
-public enum Shift {
-    MORNING("8am-11am"),
-    AFTERNOON("12pm-4pm"),
-    NIGHT("4pm-9pm");
-
-    private final String timeRange;
-
-    Shift(String timeRange) {
-        this.timeRange = timeRange;
-    }
-}
