--- src/main/java/enums/Specialty.java
@@ -1,11 +1,0 @@
-package enums;
-
-public enum Specialty {
-    DENTAL, // Deals with teeth, gums, and oral health.
-    GYNECOLOGY, // Focuses on female reproductive health and childbirth.
-    REHABILITATION, // Provides physical therapy and recovery support after injury or surgery.
-    SURGERY, // Performs operations to treat injuries, diseases, or deformities.
-    PEDIATRICS, // Provides medical care for infants, children, and adolescents.
-    RADIOLOGY, // Uses medical imaging (X-rays, MRI, CT scans) to diagnose and treat diseases.
-
-}
