--- src/main/java/enums/Status.java
@@ -1,10 +1,0 @@
-package enums;
-
-public enum Status {
-    ONLEAVE,   // The staff member is temporarily away (e.g., vacation, sick leave, maternity leave).
-    DISMISSED, // The staff member's employment was terminated by the hospital.
-    RETIRED,   // The staff member has permanently left work due to retirement.
-    ACTIVE,
-    OCCUPIED
-
-}
