+++ src/main/java/org/pahappa/systems/beans/NavigationBean.java
@@ -1,0 +1,33 @@
+package org.pahappa.systems.beans;
+
+import jakarta.enterprise.context.ApplicationScoped;
+import jakarta.inject.Named;
+
+@Named
+@ApplicationScoped
+public class NavigationBean {
+
+    public String toDashboard() {
+        return "/dashboard.xhtml?faces-redirect=true";
+    }
+
+    public String toCalendar() {
+        return "/calendar.xhtml?faces-redirect=true";
+    }
+
+    public String toPatients() {
+        return "/patients.xhtml?faces-redirect=true";
+    }
+
+    public String toAppointments() {
+        return "/appointments.xhtml?faces-redirect=true";
+    }
+
+    public String toPayments() {
+        return "/payments.xhtml?faces-redirect=true";
+    }
+
+    public String toEmployees() {
+        return "/employees.xhtml?faces-redirect=true";
+    }
+}
