+++ src/main/java/org/pahappa/systems/enums/Department.java
@@ -1,0 +1,9 @@
+package org.pahappa.systems.enums;
+
+public enum Department {
+    FINANCE,         // Billing, accounts, and financial management.
+    RECORDS,         // Medical records and health information management.
+    PHARMACY,
+    WARD
+}
+//setters
