+++ src/main/java/org/pahappa/systems/enums/Gender.java
@@ -1,0 +1,7 @@
+package org.pahappa.systems.enums;
+
+public enum Gender {
+    MALE,
+    FEMALE
+
+}
