+++ src/main/java/org/pahappa/systems/enums/PatientType.java
@@ -1,0 +1,4 @@
+package org.pahappa.systems.enums;
+
+public enum PatientType {
+}
