+++ src/main/java/org/pahappa/systems/enums/Qualification.java
@@ -1,0 +1,8 @@
+package org.pahappa.systems.enums;
+
+public enum Qualification {
+    MBBS,            // Bachelor of Medicine, Bachelor of Surgery
+    MD,              // Doctor of Medicine (postgraduate)
+    MS,              // Master of Surgery (postgraduate)
+    BDS           // Bachelor of Dental Surgery
+}
