+++ src/main/java/org/pahappa/systems/enums/Rolename.java
@@ -1,0 +1,9 @@
+package org.pahappa.systems.enums;
+
+public enum Rolename {
+    DOCTOR,
+    PHARMACIST,
+    RECEPTIONIST,
+    ADMINISTRATOR,
+    PATIENT
+}
