+++ src/main/java/org/pahappa/systems/enums/Status.java
@@ -1,0 +1,10 @@
+package org.pahappa.systems.enums;
+
+public enum Status {
+    ONLEAVE,   // The staff member is temporarily away (e.g., vacation, sick leave, maternity leave).
+    DISMISSED, // The staff member's employment was terminated by the hospital.
+    RETIRED,   // The staff member has permanently left work due to retirement.
+    ACTIVE,
+    OCCUPIED
+
+}
