+++ src/main/java/org/pahappa/systems/models/Patient.java
@@ -1,0 +1,49 @@
+package org.pahappa.systems.models;
+
+import jakarta.persistence.Entity;
+
+@Entity
+public class Patient extends User{
+private PatientType  patientType;
+private  String medicalHistory;
+private String insuranceNumber;
+
+    public Patient(int i, String janaBrincker, String s, String s1, String influenza, String s2) {
+    }
+
+    public Patient() {
+
+    }
+
+    public String getInsuranceNumber() {
+        return insuranceNumber;
+    }
+
+    public void setInsuranceNumber(String insuranceNumber) {
+        this.insuranceNumber = insuranceNumber;
+    }
+
+    public String getMedicalHistory() {
+        return medicalHistory;
+    }
+
+    public void setMedicalHistory(String medicalHistory) {
+        this.medicalHistory = medicalHistory;
+    }
+    @Override
+    public String toString() {
+        return 
+              "ID: " + getId() + "\n"
+                + "Name: " + getFullName() + "\n"
+                + "Phone Number: " + getPhoneNumber() + "\n"
+                + "Email: " + getEmail() + "\n"
+                + "Date of Birth: " + getDateOfBirth() + "\n"
+                + "Address: " + getAddress() + "\n"
+                + "Next of Kin: " + getNextOfKin() + "\n"
+                + "Role: " + getRole() + "\n"
+                + "Gender: " + getGender() + "\n"
+                + "Medical History: " + medicalHistory + "\n"
+                + "Insurance Number: " + insuranceNumber + "\n"
+                ;
+    }
+}
