--- src/main/java/repository/AppointmentDAO.java
@@ -1,83 +1,0 @@
-package repository;
-
-import models.Appointment;
-import models.Patient;
-import util.HibernateUtil;
-import org.hibernate.Session;
-import org.hibernate.Transaction;
-
-import java.util.List;
-
-public class AppointmentDAO {
-    public void saveAppointment(Appointment appointment) {
-        Transaction transaction = null;
-        try (Session session = HibernateUtil.getSessionFactory().openSession()) {
-            transaction = session.beginTransaction();
-            session.persist(appointment);
-            transaction.commit();
-        } catch (Exception e) {
-            if (transaction != null) {
-                transaction.rollback();
-            }
-            e.printStackTrace();
-        }
-    }
-
-    public List<Appointment> getAllAppointments() {
-        try (Session session = HibernateUtil.getSessionFactory().openSession()) {
-            return session.createQuery("from Appointment", Appointment.class).list();
-        }
-    }
-
-    public List<Appointment> getAppointmentsByDoctorId(int doctorId) {
-        try (Session session = HibernateUtil.getSessionFactory().openSession()) {
-            return session.createQuery("from Appointment where doctor.id = :doctorId", Appointment.class)
-                    .setParameter("doctorId", doctorId)
-                    .list();
-        }
-    }
-
-    public List<Appointment> getAppointmentsByPatientId(int patientId) {
-        try (Session session = HibernateUtil.getSessionFactory().openSession()) {
-            return session.createQuery("from Appointment where patient.id = :patientId", Appointment.class)
-                    .setParameter("patientId", patientId)
-                    .list();
-        }
-    }
-
-    public List<Appointment> getAppointmentsByPatient(Patient patient) {
-        try (Session session = HibernateUtil.getSessionFactory().openSession()) {
-            return session.createQuery("from Appointment where patient.id = :patientId", Appointment.class)
-                    .setParameter("patientId", patient.getId())
-                    .list();
-        }
-    }
-
-    public void updateAppointment(Appointment appointment) {
-        Transaction transaction = null;
-        try (Session session = HibernateUtil.getSessionFactory().openSession()) {
-            transaction = session.beginTransaction();
-            session.merge(appointment);
-            transaction.commit();
-        } catch (Exception e) {
-            if (transaction != null) {
-                transaction.rollback();
-            }
-            e.printStackTrace();
-        }
-    }
-
-    public void deleteAppointment(Appointment appointment) {
-        Transaction transaction = null;
-        try (Session session = HibernateUtil.getSessionFactory().openSession()) {
-            transaction = session.beginTransaction();
-            session.remove(session.contains(appointment) ? appointment : session.merge(appointment));
-            transaction.commit();
-        } catch (Exception e) {
-            if (transaction != null) {
-                transaction.rollback();
-            }
-            e.printStackTrace();
-        }
-    }
-}
