--- src/main/java/repository/UserDAO.java
@@ -1,116 +1,0 @@
-package repository;
-
-import enums.Specialty;
-import enums.Status;
-import models.Doctor;
-import models.User;
-import util.HibernateUtil;
-import org.hibernate.Session;
-import org.hibernate.Transaction;
-
-import java.util.List;
-
-public class UserDAO {
-    public void saveRecord(User user) {
-        Transaction transaction = null;
-        try (Session session = HibernateUtil.getSessionFactory().openSession()) {
-            transaction = session.beginTransaction();
-            session.persist(user);
-            transaction.commit();
-        } catch (Exception e) {
-            if (transaction != null) {
-                transaction.rollback();
-            }
-            e.printStackTrace();
-        }
-    }
-
-    public void updateRecord(User user) {
-        Transaction transaction = null;
-        try (Session session = HibernateUtil.getSessionFactory().openSession()) {
-            transaction = session.beginTransaction();
-            session.merge(user);
-            transaction.commit();
-        } catch (Exception e) {
-            if (transaction != null) {
-                transaction.rollback();
-            }
-            e.printStackTrace();
-        }
-    }
-
-    public User getRecordById(int id) {
-        try (Session session = HibernateUtil.getSessionFactory().openSession()) {
-            return session.get(User.class, id);
-        }
-    }
-
-    public User getRecordByName(String fullName) {
-        try (Session session = HibernateUtil.getSessionFactory().openSession()) {
-            return session.createQuery("FROM User WHERE fullName = :name", User.class)
-                    .setParameter("name", fullName)
-                    .uniqueResult();
-        }
-    }
-
-    public User getRecordByEmail(String email) {
-        try (Session session = HibernateUtil.getSessionFactory().openSession()) {
-            // return session.get(User.class, email);
-            return session.createQuery("FROM User WHERE email = :email", User.class)
-                    .setParameter("email", email)
-                    .uniqueResult();
-        }
-    }
-
-    public List<User> getAllRecords() {
-        try (Session session = HibernateUtil.getSessionFactory().openSession()) {
-            return session.createQuery("from User", User.class).list();
-        }
-    }
-
-    public void deleteRecord(int id) {
-        Transaction transaction = null;
-        try (Session session = HibernateUtil.getSessionFactory().openSession()) {
-            transaction = session.beginTransaction();
-            User user = session.get(User.class, id);
-            if (user != null) {
-                session.delete(user);
-            //    System.out.println("Student is deleted");
-            }
-            transaction.commit();
-        } catch (Exception e) {
-            if (transaction != null) {
-                transaction.rollback();
-            }
-            e.printStackTrace();
-        }
-    }
-
-    public void deleteUser(User user) {
-        Transaction transaction = null;
-        try (Session session = HibernateUtil.getSessionFactory().openSession()) {
-            transaction = session.beginTransaction();
-            if (user != null) {
-                session.delete(user);
-                System.out.println("Student is deleted");
-            }
-            transaction.commit();
-        } catch (Exception e) {
-            if (transaction != null) {
-                transaction.rollback();
-            }
-            e.printStackTrace();
-        }
-    }
-
-    public List<Doctor> getDoctorsBySpecializationAndStatus(Specialty specialization, Status status) {
-        try (Session session = HibernateUtil.getSessionFactory().openSession()) {
-            return session
-                    .createQuery("FROM Doctor WHERE specialization = :specialization AND staffStatus = :status",
-                            Doctor.class)
-                    .setParameter("specialization", specialization)
-                    .setParameter("status", status)
-                    .list();
-        }
-    }
-}
