--- src/main/java/services/UserService.java
@@ -1,12 +1,0 @@
-package services;
-
-import models.User;
-
-public interface UserService {
-    User user = null;
-   default  User login(String email, String password){
-        return user;
-   };
-
-    default void register(User user){};
-}
