+++ src/main/resources/messages.properties
@@ -1,0 +1,1 @@
+javax.faces.validator.RegexValidator.NOT_MATCHED=Email must contain @ and .com
