+++ src/main/webapp/login.xhtml
@@ -1,41 +1,130 @@
 <!DOCTYPE html>
-<html xmlns="http://www.w3.org/1999/xhtml"
-      xmlns:h="http://xmlns.jcp.org/jsf/html"
-      xmlns:p="http://primefaces.org/ui"
-      xmlns:f="http://xmlns.jcp.org/jsf/core">
+<html xmlns="http://www.w3.org/1999/xhtml" xmlns:h="http://xmlns.jcp.org/jsf/html"
+    xmlns:f="http://xmlns.jcp.org/jsf/core" xmlns:p="http://primefaces.org/ui"
+    xmlns:pt="http://xmlns.jcp.org/jsf/passthrough">
+
 <h:head>
+    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Login - Citycare General Hospital</title>
-    <h:outputStylesheet name="css/styles.css"/>
+
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet" />
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.10.5/font/bootstrap-icons.css" />
+    <style>
+        body {
+            min-height: 100vh;
+            background: #eaf1fb;
+            overflow: hidden;
+        }
+
+        .login-container {
+            min-height: 100vh;
+        }
+
+        .login-image {
+            background-size: cover;
+            min-height: 100vh;
+        }
+
+        .login-form {
+            display: flex;
+            align-items: center;
+            justify-content: center;
+            min-height: 100vh;
+            background: #fff;
+        }
+
+        .form-control:focus {
+            border-color: #0d6efd;
+            box-shadow: 0 0 0 0.2rem rgba(13, 110, 253, .25);
+        }
+
+        .btn-primary {
+            background-color: #0d6efd;
+            border: none;
+        }
+
+        .btn-primary:hover {
+            background-color: #0b5ed7;
+        }
+    </style>
+
 </h:head>
-<h:body styleClass="login-body">
-    <h:form id="loginForm">
-        <p:panel styleClass="login-panel" header="Citycare General Hospital">
-            <h:outputText value="Login into your account" styleClass="sub-header"/>
 
-            <p:messages id="messages" showDetail="true" autoUpdate="true"/>
+<h:body>
+    <h:graphicImage name="images/login_background.jpg" style="display:none" id="loginBgImg" />
+    <div class="container login-container py-5">
+        <div class="row g-0 align-items-center justify-content-center" style="min-height: 80vh;">
+            <div class="col-lg-6 d-none d-lg-block">
+                <div class="login-image rounded-start" id="loginImageDiv">
+                </div>
+            </div>
+            <div class="col-lg-6 login-form">
+                <div class="container w-75 mx-auto">
+                    <h2 class="mb-4 text-primary text-center">Welcome back </h2>
 
-            <p:outputLabel for="email" value="Email address" />
-            <p:inputText id="email" value="#{loginBean.email}" required="true" styleClass="input-full">
-                <f:facet name="prepend">
-                    <i class="pi pi-envelope"></i>
-                </f:facet>
-            </p:inputText>
 
-            <p:outputLabel for="password" value="Password" />
-            <p:password id="password" value="#{loginBean.password}" required="true" feedback="false" styleClass="input-full">
-                <f:facet name="prepend">
-                    <i class="pi pi-lock"></i>
-                </f:facet>
-            </p:password>
+                    <h:form>
+                        <h:messages globalOnly="true" style="color:red;" />
+                        <div class="mb-4">
+                            <div class="input-group">
+                                <span class="input-group-text bg-transparent border-0 border-bottom rounded-0 px-0"
+                                    id="email-addon"> </span>
+                                <h:inputText value="#{loginBean.email}"
+                                    styleClass="form-control border-0 border-bottom rounded-0 shadow-none px-0"
+                                    style="background: transparent;" id="email" aria-describedby="email-addon"
+                                    pt:placeholder="Enter email" required="true">
+<!--                                    <f:validateLength minimum="2" />-->
+<!--                                    <f:validateRegex pattern="^[^@\\s]+@[^@\\s]+\\.com$" />-->
+                                </h:inputText>
+                                <h:message for="email" styleClass="text-danger small" />
+                            </div>
+                        </div>
 
-            <h:panelGroup styleClass="forgot-password">
-                <h:outputLink value="#" onclick="PF('dlg').show(); return false;">Forgot password?</h:outputLink>
-            </h:panelGroup>
+                        <div class="mb-4">
+                            <div class="input-group">
+                                <span class="input-group-text bg-transparent border-0 border-bottom rounded-0 px-0"
+                                    id="password-addon"> </span>
+                                <h:inputSecret value="#{loginBean.password}" toggleMask="true" redisplay="true"
+                                    styleClass="form-control border-0 border-bottom rounded-0 shadow-none px-0"
+                                    style="background: transparent;" id="password" aria-describedby="password-addon"
+                                    pt:placeholder="Enter Password" required="true">
+                                    <f:validateLength minimum="2" />
+                                </h:inputSecret>
+                                <p:message for="email" display="icon" />
+                            </div>
+                        </div>
 
-            <p:commandButton value="Log in" action="#{loginBean.login}" styleClass="login-btn" ajax="false"/>
-            <div class="or-divider">OR</div>
-            <p:button value="Register" outcome="register" styleClass="register-btn"/>
-        </p:panel>
-    </h:form>
+                        <div class="d-flex justify-content-end">
+                            <h:commandButton value="Login" action="#{loginBean.login}"
+                                styleClass="btn btn-primary btn-lg px-4" />
+                        </div>
+
+                        <div class="d-flex justify-content-end mt-2">
+                            <h:outputLink value="#" styleClass="text-decoration-none small text-primary">
+                                Forgot password?
+                            </h:outputLink>
+                        </div>
+                    </h:form>
+                </div>
+            </div>
+        </div>
+    </div>
+
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"> </script>
+    <script type="text/javascript">
+        window.onload = function () {
+            var img = document.getElementById("loginBgImg");
+            var url = img ? img.src : '';
+            if (url) {
+                var div = document.getElementById('loginImageDiv');
+                if (div) {
+                    div.style.background = "url('" + url + "') no-repeat center center";
+                    div.style.backgroundSize = "cover";
+                }
+            }
+        }
+    </script>
 </h:body>
+
 </html>
