+++ src/main/webapp/resources/css/styles.css
@@ -1,57 +1,57 @@
-.login-body {
-    background: #f3faff url("background-light.svg") no-repeat center top;
-    background-size: cover;
-    font-family: 'Segoe UI', sans-serif;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    height: 100vh;
+/* Sets the light grey background for the main content area */
+body {
+    background-color: #f4f7fc;
+    overflow-x: hidden; /* Prevents horizontal scrollbar */
 }
 
-.login-panel {
-    width: 360px;
-    padding: 2rem;
-    background: white;
-    box-shadow: 0 8px 24px rgba(0, 0, 0, 0.1);
-    border-radius: 12px;
-    text-align: center;
+#wrapper {
+    display: flex; /* This is the core layout driver */
 }
 
-.sub-header {
-    font-size: 14px;
-    margin-bottom: 20px;
-    color: #666;
+/* --- Sidebar Styles --- */
+#sidebar-wrapper {
+    min-height: 100vh;
+    width: 260px; /* Fixed width for the sidebar */
+    background-color: #5d87ff; /* The specific blue from the design */
+    color: white;
+    transition: margin .25s ease-out;
 }
 
-.input-full {
-    width: 100%;
-    margin-bottom: 1rem;
+.sidebar-heading {
+    padding: 1.5rem 1.25rem;
+    font-size: 1.2rem;
+    font-weight: bold;
 }
 
-.login-btn {
-    width: 100%;
-    background-color: #007bff;
+#sidebar-wrapper .list-group-item {
+    background-color: #5d87ff; /* Match sidebar background */
+    color: #e0e0e0; /* A light grey for inactive text */
     border: none;
-    margin-top: 1rem;
+    padding: 1rem 1.5rem;
 }
 
-.register-btn {
-    width: 100%;
-    margin-top: 1rem;
-    border: 1px solid #007bff;
-    background-color: white;
-    color: #007bff;
+#sidebar-wrapper .list-group-item i {
+    margin-right: 12px;
 }
 
-.forgot-password {
-    text-align: right;
-    font-size: 12px;
-    margin-top: -10px;
-    margin-bottom: 1rem;
+/* This is the style for the selected menu item */
+#sidebar-wrapper .list-group-item.active {
+    background-color: #ffffff;
+    color: #5d87ff;
+    font-weight: bold;
+    /* This creates the curved corner on the left */
+    border-top-left-radius: 30px;
+    border-bottom-left-radius: 30px;
+    margin-right: 1rem; /* Pulls it slightly away from the edge */
 }
 
-.or-divider {
-    margin-top: 1rem;
-    color: #aaa;
-    font-size: 12px;
+#sidebar-wrapper .list-group-item:hover {
+    color: #ffffff; /* Make text white on hover */
+}
+
+
+/* --- Page Content Styles --- */
+#page-content-wrapper {
+    /* Takes up the remaining width */
+    width: calc(100% - 260px);
 }
