+++ src/main/webapp/template/template.xhtml
@@ -1,0 +1,99 @@
+<!DOCTYPE html>
+<html xmlns="http://www.w3.org/1999/xhtml"
+      xmlns:h="http://xmlns.jcp.org/jsf/html"
+      xmlns:ui="http://xmlns.jcp.org/jsf/facelets">
+<h:head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Citycare General Hospital</title>
+
+    <!-- Bootstrap CSS -->
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css" rel="stylesheet" />
+    <!-- Font Awesome for Icons -->
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.4.2/css/all.min.css" />
+
+    <!-- JSF will automatically find this in resources/css/ -->
+    <h:outputStylesheet name="css/styles.css" />
+</h:head>
+<h:body>
+
+    <!-- Main Flexbox Wrapper -->
+    <div class="d-flex" id="wrapper">
+
+        <!-- ===== 1. SIDEBAR ===== -->
+        <div class="sidebar-container" id="sidebar-wrapper">
+            <div class="sidebar-heading">
+                <i class="fa-solid fa-filter"></i> Citycare General Hopital
+            </div>
+            <div class="list-group list-group-flush">
+                <!-- 'active' class highlights the current page -->
+                <a href="#" class="list-group-item list-group-item-action active">
+                    <i class="fa-solid fa-table-columns"></i> Dashboard
+                </a>
+                <a href="#" class="list-group-item list-group-item-action">
+                    <i class="fa-solid fa-calendar-days"></i> Calendar
+                </a>
+                <a href="#" class="list-group-item list-group-item-action">
+                    <i class="fa-solid fa-hospital-user"></i> Patients
+                </a>
+                <a href="#" class="list-group-item list-group-item-action">
+                    <i class="fa-solid fa-file-medical"></i> Appointments
+                </a>
+                <a href="#" class="list-group-item list-group-item-action">
+                    <i class="fa-solid fa-credit-card"></i> Payments
+                </a>
+                <a href="#" class="list-group-item list-group-item-action">
+                    <i class="fa-solid fa-users"></i> Employees
+                </a>
+            </div>
+        </div>
+
+        <!-- ===== 2. PAGE CONTENT WRAPPER ===== -->
+        <div id="page-content-wrapper">
+
+            <!-- Top Navigation Bar -->
+            <nav class="navbar navbar-expand-lg navbar-light bg-white border-bottom">
+                <div class="container-fluid">
+                    <h:form>
+                        <div class="input-group">
+                            <h:inputText styleClass="form-control" placeholder="Search here..." />
+                        </div>
+                    </h:form>
+
+                    <ul class="navbar-nav ms-auto mt-2 mt-lg-0 align-items-center">
+                        <li class="nav-item"><a class="nav-link" href="#!"><i class="fa-solid fa-comment-medical fs-5"></i></a></li>
+                        <li class="nav-item"><a class="nav-link" href="#!"><i class="fa-solid fa-bell fs-5"></i></a></li>
+                        <li class="nav-item dropdown">
+                            <a class="nav-link dropdown-toggle d-flex align-items-center" id="navbarDropdown" href="#" role="button" data-bs-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
+                                <img src="https://i.pravatar.cc/40?u=shelia" class="rounded-circle me-2" alt="user" />
+                                <div>
+                                    <span class="fw-bold">Shelia Mugisha</span><br/>
+                                    <small class="text-muted">Receptionist</small>
+                                </div>
+                            </a>
+                            <div class="dropdown-menu dropdown-menu-end" aria-labelledby="navbarDropdown">
+                                <a class="dropdown-item" href="#!">Profile</a>
+                                <a class="dropdown-item" href="#!">Settings</a>
+                                <div class="dropdown-divider"></div>
+                                <a class="dropdown-item" href="#!">Logout</a>
+                            </div>
+                        </li>
+                    </ul>
+                </div>
+            </nav>
+
+            <!-- Main Content Area -->
+            <main class="container-fluid p-4">
+                <!-- THIS IS THE PLACEHOLDER -->
+                <ui:insert name="content">
+                    <!-- Default content can go here if a page doesn't define its own -->
+                </ui:insert>
+            </main>
+        </div>
+    </div>
+
+    <!-- JavaScript at the end of the body -->
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js"></script>
+    <script src="https://cdn.jsdelivr.net/npm/chart.js"></script>
+</h:body>
+</html>
